+++ code/packages/haskell/atbash-cipher/test/Spec.hs
@@ -1,5 +1,8 @@
 import AtbashCipherSpec (spec)
+import qualified GeneratedClassicalCipherFixtureSpec
 import Test.Hspec (hspec)
 
 main :: IO ()
-main = hspec spec
+main = hspec $ do
+    spec
+    GeneratedClassicalCipherFixtureSpec.spec